Change support tox Arch Linux

+++ src/steamconnect_game_installation.sh
@@ -1,17 +1,16 @@
 #!/bin/bash
 
-TEMP_STEAM_PASSWORD=$( dialog --stdout --title 'Confirmação' --inputbox "Por gentileza confirme sua senha da Steam" 10 50)
+TEMP_STEAM_PASSWORD=$( dialog --stdout --title 'Confirmation' --passwordbox "Please confirm your steam's password" 10 50)
 TEMP_STEAM_USERNAME=$( cat /tmp/chimera_os/steam_username.txt )
 TEMP_STEAM_GAME_ID=$( cat /tmp/chimera_os/temp_gameid.txt )
 TEMP_STEAM_GAME_NAME=$( cat /tmp/chimera_os/temp_gamename.txt )
 TEMP_STEAM_GAME_PLAT=$( cat /tmp/chimera_os/temp_platforms.txt )
 
+if [[ $TEMP_STEAM_PASSWORD == '' ]]; then
+    echo "[ERROR]: You just canceled the operation or didn't type anything"
+    exit 1
+fi
 clear
-
-#echo $TEMP_STEAM_PASSWORD
-#echo $TEMP_STEAM_USERNAME
-#echo $TEMP_STEAM_GAME_ID
-#echo $TEMP_STEAM_GAME_NAME
 
 #removing temp files
 sudo rm /tmp/chimera_os/temp_gameid.txt
